requirements.txt
# Core dependencies
requests==2.32.3
pillow==11.1.0
python-multipart==0.0.20

# FastAPI for API server
fastapi==0.115.11
uvicorn==0.34.0

# Utilities
numpy==2.2.3
pydantic==2.10.6

# Optional - for image preprocessing if needed
opencv-python==4.11.0.86

# For testing
pytest==7.4.4